+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Bug fixes:
 * Fix `StringIO#write(str)` when `str` is of an incompatible encoding and position < buffer size (#2770, @eregon).
 * Fix `rb_thread_fd_select()` to correctly initialize fdset copies and handle the timeout (@eregon).
 * Fix `TracePoint#inspect` when it's called outside of a callback (@andrykonchin).
+* Fix `Signal.trap` when signal argument is not supported (#2774, @andrykonchin).
 
 Compatibility: